changed integer field to float field

+++ erpAPP/models.py
@@ -6,17 +6,17 @@ from django.urls import reverse
 
 class fm_user_extend(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    uopeningBalance = models.IntegerField(default=0)
-    uBalance = models.IntegerField(default=0)
-    uUnconfirmed = models.IntegerField(default=0)
-    uDeclined = models.IntegerField(default=0)
+    uopeningBalance = models.FloatField(default=0)
+    uBalance = models.FloatField(default=0)
+    uUnconfirmed = models.FloatField(default=0)
+    uDeclined = models.FloatField(default=0)
 
 class fm_txn(models.Model):
     txnID = models.CharField(max_length=20,primary_key=True)
-    accID = models.IntegerField(default=0)
+    accID = models.FloatField(default=0)
     txnDate = models.DateField(auto_now=False,auto_now_add=False,null=False,blank=False)
     txnPostedDate = models.DateField(auto_now=False,auto_now_add=False,null=False,blank=False)
-    txnCheque = models.IntegerField()
+    txnCheque = models.FloatField()
     txnDir = models.CharField(max_length=2)
     txnDesc = models.CharField(max_length=100)
     txnValue = models.FloatField()
@@ -25,9 +25,9 @@ class fm_txn(models.Model):
     txnType = models.CharField(max_length=1,default='U')
     txnAccounted = models.BooleanField(default=0)
     txnAudited = models.BooleanField(default=0)
-    prID = models.IntegerField(null=True)
-    ptcID = models.IntegerField(null=True)
-    bhlID = models.IntegerField(null=True)
+    prID = models.FloatField(null=True)
+    ptcID = models.FloatField(null=True)
+    bhlID = models.FloatField(null=True)
     transc_time = models.DateTimeField(auto_now_add=True)
     def __str__(self):
         return self.txnID
@@ -40,7 +40,7 @@ class fm_utrans(models.Model):
     utranDate = models.DateField(auto_now_add=True)
     txnID = models.CharField(max_length=20,null=True)
     utranDesc = models.CharField(max_length=50,verbose_name='Description')
-    utranValue = models.IntegerField(verbose_name='Amount',null=True)
+    utranValue = models.FloatField(verbose_name='Amount',null=True)
     utranConfirmed = models.CharField(max_length=1,default='N')
     utranSender = models.CharField(max_length=10,default='no')
     utranReceiver = models.ForeignKey(User, on_delete=models.CASCADE,verbose_name='Receiver')
@@ -56,15 +56,15 @@ class fm_project(models.Model):
     prManagers_4 = models.ForeignKey(User,blank=True,null=True,on_delete=models.SET_NULL,related_name='fourth_project_manager',verbose_name ='Project Manager 4')
     prManagers_5 = models.ForeignKey(User,blank=True,null=True,on_delete=models.SET_NULL,related_name='fifth_project_manager',verbose_name ='Project Manager 5')
     prManagers_6 = models.ForeignKey(User,blank=True,null=True,on_delete=models.SET_NULL,related_name='sixth_project_manager',verbose_name ='Project Manager 6')
-    prBudget = models.IntegerField(default=0,verbose_name='Total Budget')
-    prBalance = models.IntegerField(default=0,verbose_name='Balance of budget')
+    prBudget = models.FloatField(default=0,verbose_name='Total Budget')
+    prBalance = models.FloatField(default=0,verbose_name='Balance of budget')
     def __str__(self):
         return self.prTitle
 class fm_budgethead(models.Model):
     prID = models.ForeignKey(fm_project,on_delete = models.CASCADE)
     bhTitle = models.CharField(max_length=50)
-    bhLimit = models.IntegerField(default=0)
-    bhBalance = models.IntegerField(default=0)
+    bhLimit = models.FloatField(default=0)
+    bhBalance = models.FloatField(default=0)
     bhBalanceDate = models.DateField(auto_now=True)
     def __str__(self):
         return self.bhTitle
@@ -72,7 +72,7 @@ class fm_budgethead(models.Model):
 
 class fm_ptcform(models.Model):
     uID = models.ForeignKey(User,on_delete=models.CASCADE)
-    ptcValue = models.IntegerField(default=0)
+    ptcValue = models.FloatField(default=0)
     ptcType = models.CharField(max_length=40,null=True)
     prID = models.ForeignKey(fm_project,on_delete=models.CASCADE)
     ptcDate = models.DateField(auto_now=True)
@@ -85,7 +85,7 @@ class fm_ptcform(models.Model):
 class fm_ptctrans(models.Model):
     ptctransDate = models.DateField(auto_now=False,auto_now_add=False)
     ptcID = models.ForeignKey(fm_ptcform,on_delete=models.CASCADE)
-    ptctransValue = models.IntegerField()
+    ptctransValue = models.FloatField()
     ptctransInvoiceStatus = models.CharField(max_length=1)
     ptctransInvoiceFile = models.FileField(upload_to="InvoiceFile",max_length=100,unique=True)
     ptctransHead = models.ForeignKey(fm_budgethead,on_delete=models.CASCADE)
